create useCategory composable
